MINOR: [JS] Update dev dependency gulp-vinyl-size
A minor version bump for [gulp-vinyl-size](https://www.npmjs.com/package/gulp-vinyl-size). As the first big user of that thing I thought I'd let ya'll know of the update.

The only difference for Apache Arrow will be that when `yarn run test:bundle` is run, the size values will be in SI units `kB` (1000B) instead of JEDEC `KB` (1024B).

It'll also remove the need for both gzip-size `5.1.1` _and_ `6.0.0` (the latter is already being pulled in by webpack-bundle-analyzer).

Closes #12726 from burntcustard/update-dependency-gulp-vinyl-size

Authored-by: burntcustard <burntcustard@gmail.com>
Signed-off-by: Dominik Moritz <domoritz@gmail.com>