ARROW-16099: [JS] RecordBatches that are compressed should throw an error
While https://issues.apache.org/jira/browse/ARROW-8674 is not implemented, ArrowJS will silently provide corrupt (not uncompressed) data to applications when compressed files are passed in. This happens to be the default from Python, e.g. with `df.to_feather('my_file')`.

Opening as a draft as I didn't find how to add a sample file for testing that this behaves as expected. I'd appreciate any pointers on which part of the test suite we can assert this behavior in.

Closes #12718 from isaacbrodsky/js-recordbatch-compression-throw

Lead-authored-by: Isaac Brodsky <isaac@unfolded.ai>
Co-authored-by: Dominik Moritz <domoritz@gmail.com>
Signed-off-by: Dominik Moritz <domoritz@gmail.com>