fix: Fix TypeScript compilation in a Node.js environment (#256)
Fixes https://github.com/apache/arrow-js/issues/45

This library runs in modern browsers, but is also intended for use in
maintained versions of Node.js, which means that the TypeScript types
cannot assume `"lib": ["DOM"]`.

This change removes this `lib`, and updates the types to compile.

Given that this library runs in the browser, it feels strange to import
types from `node`, but this is no worse than the current situation,
which already contains imports from `node`.